Move framework files to separate dir before making package
