+++ clients/download/src/framework/lib/i18n.js
