+++ clients/download/src/index.js
@@ -1,6 +1,14 @@
 import {createRoot} from "react-dom/client";
-import SPSPA from "./SPSPA";
+import SPSPA from "./framework/components/SPSPA";
+import App from "./App";
 import './index.css';
 
 createRoot(document.getElementById("root"))
-    .render(<SPSPA/>);
+    .render(
+        <SPSPA
+            requireNet={true}
+            subtitleKey="pages:download_project:subtitle"
+        >
+            <App/>
+        </SPSPA>
+    );
